cleaning up pv2 ready for main
